% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCellAreaShare.R
\name{calcCellAreaShare}
\alias{calcCellAreaShare}
\title{calcCellAreaShare}
\usage{
calcCellAreaShare(x)
}
\arguments{
\item{x}{MAgPIE object in grid-cellular (67420) resolution
containing area in Mha
for which cell area share is to be calculated}
}
\value{
magpie object in cellular resolution
}
\description{
This function transforms a cellular magpie object
             with 67420 grid cells to an raster object that can
             be plotted in the chosen projection
}
\examples{
\dontrun{
calcOutput("CellAreaShare", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
