% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNaturalEarth.R
\name{downloadNaturalEarth}
\alias{downloadNaturalEarth}
\title{downloadNaturalEarth}
\usage{
downloadNaturalEarth()
}
\description{
download Natural Earth shape files
}
\examples{
\dontrun{
a <- downloadSource("NaturalEarth")
}

}
\seealso{
\code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
\author{
Felicitas Beier
}
