% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrwaterPlots-package.R
\docType{package}
\name{mrwaterplots-package}
\alias{mrwaterplots}
\alias{mrwaterplots-package}
\title{mrwaterplots: plotting library for mrwater outputs}
\description{
The mrwaterplots package plots outputs from mrwater library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrwaterPlots}
}

}
\author{
\strong{Maintainer}: Felicitas Beier \email{beier@pik-potsdam.de}

Authors:
\itemize{
  \item Jens Heinke \email{heinke@pik-potsdam.de}
  \item Patrick von Jeetze \email{vjeetze@pik-potsdam.de}
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
}

}
