% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBivariateMap.R
\name{plotBivariateMap}
\alias{plotBivariateMap}
\title{plotBivariateMap}
\usage{
plotBivariateMap(
  x,
  y,
  xlab = NULL,
  ylab = NULL,
  breaks = 4,
  projection =
    "+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  bivariatelegendstyle = 1,
  colNA = "#d9d9d9",
  outputfolder = ".\\\\",
  name = "name",
  title = "",
  ylim = c(-6500000, 8300000),
  xlim = c(-12577316, 15581284),
  outputtype = "png"
)
}
\arguments{
\item{x}{First MAgPIE object in grid-cellular (67420) resolution
to be plotted}

\item{y}{Second MAgPIE object in grid-cellular (67420) resolution
to be plotted in same map}

\item{xlab}{name of first object for legend x label}

\item{ylab}{name of second object of legend y label}

\item{breaks}{Breaks in bivariate legend}

\item{projection}{Choose projection.
Currently available options:
"+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
for EqualEarth projection;
"+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
for RobinsonProj; and
"+proj=longlat +datum=WGS84" for LatLon projection}

\item{bivariatelegendstyle}{number between 1-9 to choose pre-defined bivariate legend colors}

\item{colNA}{color for NAs
(default: "#d9d9d9" (gray))}

\item{outputfolder}{Path to which plot should be saved}

\item{name}{Title of plot
(default: "name")}

\item{title}{Plot title displayed in the plot}

\item{ylim}{y-axis limits of plot
(default: c(-6500000, 8300000))}

\item{xlim}{x-axis limits of plot
(default: c(-12577316, 15581284))}

\item{outputtype}{Output type: pdf or jpeg}
}
\value{
map as pdf
}
\description{
This function plots a raster object of halfdegree
             resolution (with 67420 grid cells)
             and saves it as PDF
}
\author{
Felicitas Beier
}
