% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{plotMap}
\usage{
plotMap(
  x,
  projection =
    "+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  outputfolder = ".\\\\",
  name = "name",
  ylim = c(-6500000, 8300000),
  xlim = c(-12577316, 15581284),
  legendcolor = c("#7f0000", "#b30000", "#d7301f", "#ef6548", "#fc8d59", "#fdbb84",
    "#fdd49e", "#fee8c8", "#fff7ec", "#045a8d"),
  colNA = "#d9d9d9",
  legendlimit = c(0, 1),
  legendbreaks = seq(0, 1, 0.1),
  legendname = "legendname"
)
}
\arguments{
\item{x}{MAgPIE object in grid-cellular (67420) resolution
to be transformed to raster}

\item{projection}{Choose projection.
Currently available options:
"+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
for EqualEarth projection}

\item{outputfolder}{Path to which plot should be saved}

\item{name}{Title of plot
(default: "name")}

\item{ylim}{y-axis limits of plot
(default: c(-6500000, 8300000))}

\item{xlim}{x-axis limits of plot
(default: c(-12577316, 15581284))}

\item{legendcolor}{vector of colors}

\item{colNA}{color for NAs
(default: "#d9d9d9" (gray))}

\item{legendlimit}{vector with min and max of legend
(default: c(0, 1))}

\item{legendbreaks}{vector of legend breaks
(default: seq(0, 1, 0.1))}

\item{legendname}{legend name as character
(default: "legendname")}
}
\value{
magpie object in cellular resolution
}
\description{
This function plots a raster object of halfdegree
             resolution (with 67420 grid cells)
             and saves it as PDF
}
\author{
Felicitas Beier, Jens Heinke
}
