% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapLayered.R
\name{plotMapLayered}
\alias{plotMapLayered}
\title{plotMapLayered}
\usage{
plotMapLayered(
  x,
  projection =
    "+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  outputfolder = ".\\\\",
  name = "name",
  title = "",
  ylim = c(-6500000, 8300000),
  xlim = c(-12577316, 15581284),
  legendcolor = list(c("#dfc27d", "#bf812d", "#8c510a", "#543005"), c("#80cdc1",
    "#35978f", "#01665e", "#003c30")),
  legendtitles = c("scale1", "scale2"),
  legendname = "legendtitle (unit)",
  naObject = NULL,
  legendlimit = c(0, 1),
  legendbreaks = seq(0, 1, 0.25),
  outputtype = "png",
  minVal = NULL,
  maxVal = NULL
)
}
\arguments{
\item{x}{list of MAgPIE objects in grid-cellular (67420) resolution
to be plotted}

\item{projection}{Choose projection.
Currently available options:
"+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
for EqualEarth projection;
"+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
for RobinsonProj; and
"+proj=longlat +datum=WGS84" for LatLon projection}

\item{outputfolder}{Path to which plot should be saved}

\item{name}{Title of plot
(default: "name")}

\item{title}{Plot title displayed in the plot}

\item{ylim}{y-axis limits of plot
(default: c(-6500000, 8300000))}

\item{xlim}{x-axis limits of plot
(default: c(-12577316, 15581284))}

\item{legendcolor}{List of vector of colors of the same length as the list of objects
provided}

\item{legendtitles}{Vector for title for legendcolors.
Should have the same length as the list of legendcolor}

\item{legendname}{legend name as character
(default: "legendname")}

\item{naObject}{magpie object that defines which areas to mask out as NA
with 1 being the ones to be masked out and 0 the ones
that are transparent, i.e. show the previous plots}

\item{legendlimit}{vector with min and max of legend
(default: c(0, 1))}

\item{legendbreaks}{vector of legend breaks
(default: seq(0, 1, 0.1))}

\item{outputtype}{Output type: pdf or jpeg or png}

\item{minVal}{minimum value at which x should be chopped}

\item{maxVal}{maximum value at which x should be chopped}
}
\value{
map as pdf, jpg or png
}
\description{
This function plots several magpie objects of halfdegree
             resolution (with 67420 grid cells) as overlayed map
}
\author{
Felicitas Beier, Jens Heinke
}
