% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCellAreaShare.R
\name{toolCellAreaShare}
\alias{toolCellAreaShare}
\title{toolCellAreaShare}
\usage{
toolCellAreaShare(x)
}
\arguments{
\item{x}{MAgPIE object in grid-cellular (67420) resolution
containing area in Mha
for which cell area share is to be calculated}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the share of total grid cell area
             for a given cellular magpie object with 67420 grid cells
}
\author{
Felicitas Beier
}
