% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRasterTransform.R
\name{toolRasterTransform}
\alias{toolRasterTransform}
\title{toolRasterTransform}
\usage{
toolRasterTransform(
  x,
  projection =
    "+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
)
}
\arguments{
\item{x}{MAgPIE object in grid-cellular (67420) resolution
to be transformed to raster}

\item{projection}{Choose projection.
Currently available options:
"+proj=eqearth +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs" for EqualEarth projection}
}
\value{
magpie object in cellular resolution
}
\description{
This function transforms a cellular magpie object
             with 67420 grid cells to an raster object that can
             be plotted in the chosen projection
}
\author{
Felicitas Beier, Jens Heinke
}
